pyproject.toml
[project]
name = "newchan"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gae", email = "gaedev22@gmail.com" }
]
dependencies = [
    "slack-sdk>=3.31.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/newchan"]
